alteração dos tipos de dados

+++ teste_algoritmos.java
@@ -8,10 +8,10 @@ public class teste_algoritmos {
 	Scanner sc = new Scanner(System.in);
 	
 	System.out.println("Digite um número: ");
-	int a = sc.nextInt();
+	double a = sc.nextInt();
 	System.out.println("Digite outro número: ");
-	int b = sc.nextInt();
-	int resultado = 0;
+	double b = sc.nextInt();
+	double resultado = 0;
 	
 	if (a > b) {
 		resultado = (a - b) / a;
